Move offline orders functionality to core modules
